+++ backend/books/views.py
@@ -18,6 +18,7 @@ from .paginations import BookPagination
 from .search_filters import *
 from .utils import str2bool
 from utils.permissions import CustomBasePermission
+from .book_images import BookImageCreator
 
 from genres.models import Genre
 from purchase_orders.models import Purchase
@@ -108,7 +109,8 @@ class ISBNSearchView(APIView):
 
         return ret
 
-class ListCreateBookAPIView(ListCreateAPIView):
+
+class ListCreateBookAPIView(ListCreateAPIView, BookImageCreator):
     serializer_class = BookListAddSerializer
     permission_classes = [CustomBasePermission]
     pagination_class = BookPagination
@@ -182,38 +184,6 @@ class ListCreateBookAPIView(ListCreateAPIView):
 
         return data
     
-    def has_image_bytes(self, request):
-        return request.FILES.get('image_bytes', None) is not None
-
-    def has_image_url(self, request):
-        return request.data.get('image_url', None) is not None
-
-    def bookimage_get_and_create(self, request, isbn_13, setDefaultImage):
-        book = Book.objects.filter(isbn_13=isbn_13)
-
-        # This creates an image in static and sends a file
-        if setDefaultImage:
-            url = self.isbn_toolbox.get_default_image_url()
-        elif self.has_image_bytes(request):
-            url = self.isbn_toolbox.commit_image_raw_bytes(request, book[0].id, isbn_13)
-        elif self.has_image_url(request):
-            url = self.isbn_toolbox.commit_image_url(request, book[0].id, isbn_13)
-        else:
-            url = self.isbn_toolbox.get_default_image_url()
-
-
-        obj, created = BookImage.objects.get_or_create(
-            book_id=book[0].id,
-            defaults={'image_url': url},
-        )
-
-        # We need to patch the url if it is a get
-        if not created:
-            obj.image_url = url
-            obj.save()
-
-        return url
-
     def getOrCreateModel(self, item_list, model):
         if isinstance(item_list, str):
             item_list = item_list.split(',')
@@ -318,7 +288,7 @@ class ListCreateBookAPIView(ListCreateAPIView):
         return query_set.annotate(shelf_space=F('null_defaulted_thickness') * F('stock'))
 
 
-class RetrieveUpdateDestroyBookAPIView(RetrieveUpdateDestroyAPIView):
+class RetrieveUpdateDestroyBookAPIView(RetrieveUpdateDestroyAPIView, BookImageCreator):
     serializer_class = BookSerializer
     queryset = Book.objects.all()
     permission_classes = [CustomBasePermission]
@@ -377,37 +347,6 @@ class RetrieveUpdateDestroyBookAPIView(RetrieveUpdateDestroyAPIView):
                 data[possible_zero_field] = None
 
         return data
-
-    def has_image_bytes(self, request):
-        return request.FILES.get('image_bytes', None) is not None
-
-    def has_image_url(self, request):
-        return request.data.get('image_url', None) is not None
-
-    def bookimage_get_and_create(self, request, isbn_13, setDefaultImage):
-        book = Book.objects.filter(isbn_13=isbn_13)
-
-        # This creates an image in static and sends a file
-        if setDefaultImage:
-            url = self.isbn_toolbox.get_default_image_url()
-        elif self.has_image_bytes(request):
-            url = self.isbn_toolbox.commit_image_raw_bytes(request, book[0].id, isbn_13)
-        elif self.has_image_url(request):
-            url = self.isbn_toolbox.commit_image_url(request, book[0].id, isbn_13)
-        else:
-            url = self.isbn_toolbox.get_default_image_url()
-
-        obj, created = BookImage.objects.get_or_create(
-            book_id=book[0].id,
-            defaults={'image_url': url},
-        )
-
-        # We need to patch the url if it is a get
-        if not created:
-            obj.image_url = url
-            obj.save()
-
-        return url
 
     def destroy(self, request, *args, **kwargs):
